requirements.txt
addict==2.4.0
aiofiles==23.2.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.3.2
aws-lambda-powertools==3.3.0
blinker==1.9.0
boto3==1.35.65
botocore==1.35.65
brotlicffi @ file:///croot/brotlicffi_1736182461069/work
cachetools==6.2.0
certifi @ file:///croot/certifi_1750863846190/work/certifi
cffi==1.17.1
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
crcmod==1.7
cryptography==44.0.0
cycler==0.12.1
exceptiongroup==1.3.0
fastapi==0.112.4
ffmpy==0.6.0
filelock==3.14.0
Flask==3.1.0
fonttools==4.58.4
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.45
gmpy2 @ file:///croot/gmpy2_1738085463648/work
gradio==4.43.0
gradio_client==1.3.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==8.5.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mmcv==2.1.0
mmdet==3.3.0
mmengine==0.10.7
model-index==0.1.11
mpmath @ file:///croot/mpmath_1690848262763/work
narwhals==1.45.0
networkx @ file:///croot/networkx_1717597493534/work
numpy==1.26.4
opencv-python==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
orjson==3.10.18
oss2==2.17.0
packaging==24.2
pandas==2.3.0
pillow==10.4.0
platformdirs==4.3.8
protobuf==6.32.0
pyarrow==21.0.0
pycocotools==2.0.10
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.9.0
pydantic_core==2.23.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2023.4
PyYAML @ file:///croot/pyyaml_1728657952215/work
qrcode==8.0
referencing==0.36.2
requests==2.32.4
rich==13.4.2
rpds-py==0.26.0
ruff==0.12.2
s3transfer==0.10.3
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
shapely==2.0.7
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.38.6
streamlit==1.49.1
sympy @ file:///croot/sympy_1738108488918/work
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
terminaltables==3.1.10
toml==0.10.2
tomli==2.2.1
tomlkit==0.12.0
torch==2.5.1
torchvision==0.20.1
tornado==6.5.2
tqdm==4.65.2
triton==3.1.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3 @ file:///croot/urllib3_1750775463400/work
uvicorn==0.35.0
watchdog==6.0.0
websockets==11.0.3
Werkzeug==3.1.3
yapf==0.43.0
zipp==3.21.0
